+++ index.ts
@@ -1,14 +1,68 @@
-import "reflect-metadata"
 import "dotenv/config"
 
 import { ApolloServer } from "apollo-server"
-import { createConnection, Connection } from "typeorm"
 import * as AWS from "aws-sdk"
 
 import type { Request, Response } from "express"
 
 import { schema } from "./src/schema"
-import { entities } from "./src/entity"
+
+// Safe to say that running `now` updates `process.env.NODE_ENV` to `production`
+const __PROD__ = process.env.NODE_ENV === "production"
+const __DEV__ = process.env.NODE_ENV === "development"
+
+AWS.config.update(
+  __PROD__
+    ? {
+        region: "us-east-1",
+        // accessKeyId default can be used while using the downloadable version of DynamoDB.
+        // For security reasons, do not store AWS Credentials in your files. Use Amazon Cognito instead.
+        accessKeyId: process.env.MY_AWS_ACCESS_KEY_ID,
+        // secretAccessKey default can be used while using the downloadable version of DynamoDB.
+        // For security reasons, do not store AWS Credentials in your files. Use Amazon Cognito instead.
+        secretAccessKey: process.env.MY_AWS_SECRET_ACCESS_KEY,
+      }
+    : {
+        region: "localhost",
+        accessKeyId: process.env.MY_AWS_ACCESS_KEY_ID,
+        secretAccessKey: process.env.MY_AWS_SECRET_ACCESS_KEY,
+      }
+)
+
+const options:
+  | AWS.DynamoDB.ClientConfiguration
+  | AWS.DynamoDB.DocumentClient.DocumentClientOptions = __PROD__
+  ? {
+      apiVersion: "2012-08-10",
+      region: "us-east-1",
+      endpoint: "https://dynamodb.us-east-1.amazonaws.com",
+      accessKeyId: process.env.MY_AWS_ACCESS_KEY_ID,
+      secretAccessKey: process.env.MY_AWS_SECRET_ACCESS_KEY,
+    }
+  : {
+      apiVersion: "2012-08-10",
+      region: "localhost",
+      endpoint: "http://localhost:8001",
+      /**
+       * Steps w/ NoSQL Workbench
+       * 1. Run dynamo docker container
+       * 2. Create a local connection via NoSQL Workbench
+       * 3. Find credentials
+       * if connecting to docker DB, leave empty
+       */
+      accessKeyId: process.env.NO_SQL_WORKBENCH_ACCESS_KEY_ID,
+      secretAccessKey: process.env.NO_SQL_WORKBENCH_SECRET_ACCESS_KEY,
+    }
+/**
+ * Locking the API Version
+ * https://docs.aws.amazon.com/AWSJavaScriptSDK/latest/AWS/DynamoDB.html
+ */
+const dynamoDb = new AWS.DynamoDB(options)
+/**
+ * Difference between DynamoDB & DocumentClient
+ * https://stackoverflow.com/a/57807642/9823455
+ */
+const docClient = new AWS.DynamoDB.DocumentClient(options)
 
 AWS.config.update({
   region: "us-east-1",
@@ -19,55 +73,15 @@ const s3 = new AWS.S3()
 const cognito = new AWS.CognitoIdentityServiceProvider()
 
 export interface Context {
-  connection: Connection
   s3: AWS.S3
   cognito: AWS.CognitoIdentityServiceProvider
   req: Request
   res: Response
+  dynamoDb: AWS.DynamoDB
+  docClient: AWS.DynamoDB.DocumentClient
 }
 
 async function main() {
-  const host = process.env.RDS_DB_HOST
-  const port = parseInt(process.env.RDS_DB_PORT as string)
-  const username = process.env.RDS_DB_USERNAME
-  const password = process.env.RDS_DB_PASSWORD
-  const database = process.env.RDS_DB_DATABASE
-
-  const makeTypeORMConnection = async () => {
-    let connection = await createConnection({
-      name: "default",
-      type: "postgres",
-      host,
-      port,
-      username,
-      password,
-      database,
-      // synchronize: true,
-      logging: false,
-      entities: entities,
-      migrations: ["src/migration/**/*.ts"],
-      subscribers: ["src/subscriber/**/*.ts"],
-      cli: {
-        entitiesDir: "src/entity",
-        migrationsDir: "src/migration",
-        subscribersDir: "src/subscriber",
-      },
-    })
-
-    if (connection.isConnected) {
-      console.log("✔ Connected to RDS")
-      console.log(
-        "Entity names:",
-        connection.entityMetadatas.map(e => e.tableName)
-      )
-    }
-    return connection
-  }
-
-  const [connectionRes] = await Promise.allSettled([makeTypeORMConnection()])
-
-  const connection = connectionRes.status === "fulfilled" && connectionRes.value
-
   const server = new ApolloServer({
     schema: schema,
     introspection: true,
@@ -87,12 +101,19 @@ async function main() {
       ctx.res.header("Set-Cookie", "wtf=hellooooo")
       return {
         ...ctx,
-        connection,
         cognito,
+        dynamoDb,
+        docClient,
         s3,
       } as Context
     },
   })
+
+  if (__DEV__) {
+    console.log("ℹ️ You are in __DEV__ mode")
+    const tables = await dynamoDb.listTables().promise()
+    console.log(tables)
+  }
 
   server.listen({ port: 4044 }).then(({ url, subscriptionsUrl }) => {
     console.log(`🚀 Server ready at ${url}`)