+++ src/modules/User/Mutation/getOrCreateUser.ts
@@ -1,7 +1,6 @@
-import { AuthenticationError } from "apollo-server"
 import type { ResolverFn } from "../../resolverFn"
 import { decodeBearerToken } from "../../../utils"
-import { User } from "../../../entity/User"
+import { ApolloError } from "apollo-server"
 
 /**
  * Attempt to find a user with the UUID on the Bearer token
@@ -12,42 +11,9 @@ export const getOrCreateUser: ResolverFn = async function (
   context,
   info
 ) {
-  let cognitoUuid: string
+  console.log("Mutation.getOrCreateUser")
+  const decoded = await decodeBearerToken(context)
+  const id = decoded.username
 
-  try {
-    const decoded = await decodeBearerToken(context)
-    cognitoUuid = decoded.username
-  } catch (err) {
-    throw new AuthenticationError(err.toString())
-  }
-
-  console.log("searching")
-  const { connection } = context
-  const repo = connection.getRepository(User)
-  try {
-    const user = await repo
-      .createQueryBuilder("user")
-      .select()
-      .where("user.cognito_sub = :uuid", {
-        uuid: cognitoUuid,
-      })
-      .getOne()
-
-    if (!user) {
-      let user = new User()
-      user.id = cognitoUuid
-      user.cognito_sub = cognitoUuid
-      user.email = email
-      user.first_name = firstName
-      user.last_name = lastName
-      user = await connection.manager.save(user)
-      console.log("new", user)
-      return user
-    } else {
-      console.log("found", user)
-      return user
-    }
-  } catch (err) {
-    console.log(err)
-  }
+  throw new ApolloError("Unimplemented")
 }