+++ src/modules/User/Mutation/index.ts
@@ -1,1 +1,2 @@
 export { getOrCreateUser } from "./getOrCreateUser"
+export { updateUsername } from "./updateUsername"