+++ src/modules/User/Query/getUsers.ts
@@ -1,5 +1,6 @@
+import { ApolloError } from "apollo-server"
+import { DocumentClient } from "aws-sdk/clients/dynamodb"
 import type { ResolverFn } from "../../resolverFn"
-import { User } from "../../../entity/User"
 
 export const getUsers: ResolverFn = async function (
   parent,
@@ -7,12 +8,26 @@ export const getUsers: ResolverFn = async function (
   context,
   info
 ) {
-  const { connection } = context
-  const result = await connection
-    .getRepository(User)
-    .createQueryBuilder("user")
-    .limit(10)
-    .orderBy("user.created", "DESC")
-    .getMany()
-  return result
+  console.log("Query.getUsers")
+  const { docClient } = context
+
+  const params: DocumentClient.QueryInput = {
+    TableName: process.env.TABLE_NAME,
+    IndexName: "GSI_InvertedIndex",
+    KeyConditionExpression: "#8f150 = :8f150",
+    ExpressionAttributeValues: {
+      ":8f150": "#IDENTITY",
+    },
+    ExpressionAttributeNames: {
+      "#8f150": "SK",
+    },
+    Limit: 10,
+  }
+  try {
+    const res = await docClient.query(params).promise()
+    return res.Items
+  } catch (err) {
+    console.log("❌", err)
+    throw new ApolloError(err.message)
+  }
 }