+++ src/modules/User/Query/index.ts
@@ -1,1 +1,2 @@
 export { getUsers } from "./getUsers"
+export { getOrCreateUser } from "./getOrCreateUser"