+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.5.0] - 2020-12-01
+
+### Bye bye Postgres; Hello Dynamo
+
+I'm done using TypeORM and Postgres.
+
+### Added
+
+- `process.env.TABLE_NAME`
+  - this needs to be added in Vercel
+- graphql-codegen
+- update `updateUsername` to Dynamo logic
+
 ## [v0.4.0] - 2020-11-26
 
 ### BREAKING CHANGES
@@ -181,6 +194,7 @@ add skip/take to getCommentsByUrl
 
 ### Added
 
+[v0.5.0]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.4.0...v0.5.0
 [v0.4.0]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.3.1...v0.4.0
 [v0.3.1]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.3.0...v0.3.1
 [v0.3.0]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.2.0...v0.3.0